+++ samplecom/src/main/java/com/example/action/model/base/MsyKbnDeleteAction.java
@@ -33,15 +33,10 @@ public class MsyKbnDeleteAction extends BaseAction {
 
         MsyKbn e = FormValidator.toBean(MsyKbn.class.getName(), form);
 
-        // 子：区分値マスタの削除
-        java.util.List<com.example.entity.MsyKbnVal> msyKbnVals = e.referMsyKbnVals();
-        if (msyKbnVals != null) {
-            for (com.example.entity.MsyKbnVal msyKbnVal : msyKbnVals) {
-
-                if (msyKbnVal.delete() != 1) {
-                    throw new OptLockError("error.cant.delete", "区分値マスタ");
-                }
-            }
+        // 子：区分値マスタを全て指定済みか確認
+        int orgMsyKbnValsSize = e.getMsyKbnVals().size();
+        if (e.referMsyKbnVals().size() != orgMsyKbnValsSize) {
+            throw new OptLockError("error.cant.delete", "区分値マスタ");
         }
 
         if (e.delete() != 1) {
