+++ samplecom/src/main/java/com/example/action/model/base/T02OyaDeleteAction.java
@@ -33,28 +33,19 @@ public class T02OyaDeleteAction extends BaseAction {
 
         T02Oya e = FormValidator.toBean(T02Oya.class.getName(), form);
 
-        // 子：子なしの削除
-        java.util.List<com.example.entity.T02Dinks> t02Dinkss = e.referT02Dinkss();
-        if (t02Dinkss != null) {
-            for (com.example.entity.T02Dinks t02Dinks : t02Dinkss) {
-
-                if (t02Dinks.delete() != 1) {
-                    throw new OptLockError("error.cant.delete", "子なし");
-                }
-            }
+        // 子：子なしを全て指定済みか確認
+        int orgT02DinkssSize = e.getT02Dinkss().size();
+        if (e.referT02Dinkss().size() != orgT02DinkssSize) {
+            throw new OptLockError("error.cant.delete", "子なし");
         }
 
-        // 子：子の削除
-        java.util.List<com.example.entity.T02Ko> t02Kos = e.referT02Kos();
-        if (t02Kos != null) {
-            for (com.example.entity.T02Ko t02Ko : t02Kos) {
-
-                // child:T02Mago, parents:2
-
-                if (t02Ko.delete() != 1) {
-                    throw new OptLockError("error.cant.delete", "子");
-                }
-            }
+        // 子：子を全て指定済みか確認
+        int orgT02KosSize = e.getT02Kos().size();
+        if (e.referT02Kos().size() != orgT02KosSize) {
+            throw new OptLockError("error.cant.delete", "子");
+        }
+        for (com.example.entity.T02Ko t02Ko : e.getT02Kos()) {
+            t02Ko.nestT02Magos();
         }
 
         if (e.delete() != 1) {
