+++ samplecom/src/main/java/com/example/action/model/base/T07Derive1SDeleteAction.java
@@ -43,15 +43,10 @@ public class T07Derive1SDeleteAction extends BaseAction {
 
                 T07Derive1 e = FormValidator.toBean(T07Derive1.class.getName(), row);
 
-                // 子：派生１明細の削除
-                java.util.List<com.example.entity.T07Derive1Det> t07Derive1Dets = e.referT07Derive1Dets();
-                if (t07Derive1Dets != null) {
-                    for (com.example.entity.T07Derive1Det t07Derive1Det : t07Derive1Dets) {
-
-                        if (t07Derive1Det.delete() != 1) {
-                            throw new OptLockError("error.cant.delete", "派生１明細");
-                        }
-                    }
+                // 子：派生１明細を全て指定済みか確認
+                int orgT07Derive1DetsSize = e.getT07Derive1Dets().size();
+                if (e.referT07Derive1Dets().size() != orgT07Derive1DetsSize) {
+                    throw new OptLockError("error.cant.delete", "派生１明細");
                 }
 
                 if (e.delete() != 1) {
